+++ examples/wallet.mjs
@@ -2,11 +2,16 @@
 
 import { Wallet } from '../src';
 
-const wallet = new Wallet();
+const wallet = new Wallet({
+  username: 'testusername',
+  password: 'testpassword',
+});
 
 (async () => {
   try {
-    const response = await wallet.getHeight();
+    await wallet.create('testwallet', 'testpassword');
+    await wallet.open('testwallet', 'testpassword');
+    const response = await wallet.getAddress();
     console.log(response);
   } catch (err) {
     console.error(err);